+++ braindecode/models/__init__.py
@@ -33,10 +33,15 @@ from .eegnex import EEGNeX
 from .tsinception import TSceptionV1
 from .eegtcnet import EEGTCNet
 from .syncnet import SyncNet
+from .lmda import LMDANet
+from .sccnet import SCCNet
 from .fbcnet import FBCNet
+from .ifnet import IFNetV2
 from .eegminer import EEGMiner
 from .fblightconvnet import FBLightConvNet
+from .fbmsnet import FBMSNet
 from .ctnet import CTNet
+
 
 # Call this last in order to make sure the dataset list is populated with
 # the models imported in this file.